# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: xiaoan-platform API
  version: 2.0.0
  description: |
    xiaoan-platform backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/v2/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List openable applications (bare array, no pagination envelope)
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, task, custom, all], default: chat }
          description: Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [accessible, public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: Truthy when case-insensitively in {1,true,yes}.
      responses:
        "200":
          description: Application list items (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApplicationListItem" }
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application (admin only — agent marketplace "new Aily agent")
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationCreateRequest" }
      responses:
        "201":
          description: Created (compact detail shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "403":
          description: only admins may add agents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/page:
    get:
      operationId: listApplicationPage
      tags: [applications]
      summary: Keyset-paginated application catalog page (执行报告 §14–§17)
      description: |
        True server-side pagination: LIMIT / cursor-condition evaluate inside
        SQL, one anti-joined query per page (no N+1), and personal usage /
        favorites aggregate ONLY the ids on the returned page. The legacy
        GET /api/v2/applications keeps its whole-catalog array shape for
        callers not yet migrated; new surfaces must use this endpoint.
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, fixed, all], default: chat }
          description: >
            `fixed` means kind <> 'chat' (应用中心: pages/forms/dashboards/tasks).
            Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [accessible, public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: mode
          in: query
          schema: { type: string, enum: [manage, consume], default: manage }
          description: >
            WHO may see the row is `scope`; whether the row may actually be
            USED is `mode` (二次复审 P0-5). `manage` (default) is the
            management surface (智能体市场 / 应用中心): staff also sees
            disabled, private and binding-less rows so they can repair them.
            `consume` is every surface that opens or runs something
            (切换器 / 移动端目录 / bootstrap 分组): it requires
            `enabled = 1` and, for `kind = chat`, an enabled runtime binding
            whose provider exists and is active — exactly the predicate
            `AuthorizeExecution` enforces, so the UI can no longer offer an
            application that the run API then refuses.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: >
            Truthy when case-insensitively in {1,true,yes}; keeps binding-less
            CHAT rows (the management chat catalog's knob). It no longer
            affects `fixed` rows in any mode: a fixed application never needs
            a binding to be consumable, so `kind=fixed&mode=consume` returns
            unbound fixed applications whether or not this flag is set
            (三次复审 P1-R4).
        - name: q
          in: query
          schema: { type: string }
          description: Case-insensitive substring match on name / description / category name.
        - name: category_slug
          in: query
          schema: { type: string }
          description: >
            Exact category match; the sentinel `__uncategorized__` matches rows
            without any category (mobile 其他 tab).
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
        - name: cursor
          in: query
          schema: { type: string }
          description: Opaque next_cursor from a previous response (base64 JSON of created_at+id).
      responses:
        "200":
          description: One page of application list items (same item shape as the legacy array)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationPage" }
        "400":
          description: invalid cursor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/resolve:
    get:
      operationId: resolveApplication
      tags: [applications]
      summary: Resolve ONE application by slug or id (执行报告 §12)
      description: >
        Exact single-row resolution for deep links (/chat/:slug, /app/:slug) and
        for the application entity cache. Visibility is evaluated with the
        enterprise access policy (legacy is_public while the rollout flag is off;
        Directory + Application ACL when enabled; staff bypass ACL); then consumption eligibility is applied: the application
        must be enabled and a chat application must have an enabled binding
        backed by an active provider. It answers 404 when the application does
        not exist OR is not visible/consumable — existence is never leaked. This is what makes
        "open /chat/sales" cost one small request instead of a catalog
        download plus Array.find().
      parameters:
        - name: slug
          in: query
          schema: { type: string }
          description: Exact slug. Provide this or `id`.
        - name: id
          in: query
          schema: { type: integer }
          description: Exact application id. Provide this or `slug`.
      responses:
        "200":
          description: The application, in the same shape the catalog page uses
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationListItem" }
        "400":
          description: neither slug nor id was provided
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "404":
          description: unknown, not visible, or not usable by this caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/resolve-mention:
    get:
      operationId: resolveApplicationMention
      tags: [applications]
      summary: Resolve the `@` token of a composer instruction to candidates (§14)
      description: >
        Server-side replacement for the client-side @mention candidate array
        (which was built by downloading every application). Matching is
        case-insensitive on the application name and slug; the response is
        ranked (exact name/slug first, then name prefix, then slug prefix,
        then substring) and capped, so the composer never needs the whole catalog
        to route `@销售助手 帮我查一下`.

        Like `resolve`, this is a CONSUMPTION entry point (二次复审 P0-5): only
        usable applications are candidates, so `@` can never switch the
        composer to an application the next send would be refused for.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: The text after `@`, e.g. `销售助手`.
      responses:
        "200":
          description: Ranked candidates (at most 10)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentionCandidate" }
        "400":
          description: q is required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
components:
...
  schemas:
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
...
    FieldErrors:
      type: object
      description: map of field name -> list of messages
      additionalProperties:
        type: array
        items: { type: string }
...
    ApplicationListItem:
      type: object
      description: shape returned by GET /api/v2/applications
      required:
        - id
        - slug
        - name
        - description
        - icon
        - avatar_url
        - color
        - kind
        - renderer_key
        - executor_key
        - category_slug
        - category_name
        - is_public
        - enabled
        - runtime_type
        - provider_key
        - external_resource_id
        - identity_mode
        - execution_mode
        - capabilities
        - is_bound
        - is_consumable
        - is_favorite
        - is_default_agent
        - can_manage
        - usage_count
        - last_used_at
        - global_usage_count
        - skills
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        avatar_url: { type: string }
        color: { type: string }
        kind: { type: string, enum: [chat, task, custom] }
        renderer_key: { type: string }
        executor_key: { type: string }
        category_slug: { type: string }
        category_name: { type: string }
        is_public: { type: boolean }
        enabled: { type: boolean }
        runtime_type: { type: string }
        provider_key: { type: string }
        external_resource_id: { type: string }
        identity_mode: { type: string }
        execution_mode: { type: string }
        capabilities:
          type: object
          additionalProperties: { type: boolean }
        is_bound: { type: boolean }
        is_consumable: { type: boolean }
        consume_block_reason:
          type: string
          enum: [disabled, unbound, runtime_unavailable]
        is_favorite: { type: boolean }
        is_default_agent: { type: boolean }
        can_manage: { type: boolean }
        usage_count: { type: integer }
        last_used_at:
          type: string
          nullable: true
          format: date-time
        global_usage_count: { type: integer }
        skills:
          type: array
          description: >-
            Agent-scoped 技能配置. Always present; `[]` when the agent has none.
          items: { $ref: "#/components/schemas/ApplicationSkill" }
    ApplicationPage:
      type: object
      description: >
        One keyset page of the catalog (GET /api/v2/applications/page). Items
        carry the same shape as the legacy bare-array endpoint. next_cursor is
        opaque and empty when has_more is false.
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ApplicationListItem" }
        next_cursor: { type: string }
        has_more: { type: boolean }
...
    MentionCandidate:
      type: object
      description: >
        One `@` candidate for the composer's routing decision. `kind` matters:
        a chat application receives the instruction, anything else is opened.
      required: [id, slug, name, kind]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        kind: { type: string, enum: [chat, task, custom] }
    ApplicationSkill:
      type: object
      description: >-
        An agent-scoped skill: a NAMED prompt fragment the composer prepends to
        the user's message. It is not a provider capability — the provider only
        ever sees ordinary message content, and interpreting the fragment is the
        custom agent's own concern. Because the fragment changes `content`, the
        existing idempotency request hash already covers it.
      required: [id, name, prompt]
      properties:
        id:
          type: string
          maxLength: 64
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: >-
            Stable across renames so a saved composer selection survives an
            admin rewording the skill. Derived from `name` when omitted.
        name: { type: string, maxLength: 24, description: composer chip label }
        description: { type: string, maxLength: 60, description: one-line subtitle }
        prompt:
          type: string
          maxLength: 200
          description: prepended verbatim to the user's message
    ApplicationDetail:
      type: object
      description: compact shape returned by authoring / avatar / default-agent endpoints
      required:
        - id
        - slug
        - name
        - description
        - icon
        - avatar_url
        - color
        - kind
        - is_public
        - enabled
        - category_slug
        - category_name
        - is_default_agent
        - is_bound
        - is_consumable
        - runtime_type
        - provider_key
        - external_resource_id
        - identity_mode
        - execution_mode
        - can_manage
        - updated_at
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        icon: { type: string }
        avatar_url: { type: string }
        color: { type: string }
        kind: { type: string, enum: [chat, task, custom] }
        is_public: { type: boolean }
        enabled: { type: boolean }
        category_slug: { type: string }
        category_name: { type: string }
        is_default_agent: { type: boolean }
        is_bound: { type: boolean }
        is_consumable: { type: boolean }
        consume_block_reason:
          type: string
          enum: [disabled, unbound, runtime_unavailable]
        runtime_type: { type: string }
        provider_key: { type: string }
        external_resource_id: { type: string }
        identity_mode: { type: string }
        execution_mode: { type: string }
        can_manage: { type: boolean }
        updated_at: { type: string, format: date-time }
        skills:
          type: array
          description: Agent-scoped 技能配置; always present, `[]` when none.
          items: { $ref: "#/components/schemas/ApplicationSkill" }
    RuntimeBindingInput:
      type: object
      required: [provider_key]
      properties:
        provider_key: { type: string, maxLength: 64 }
        runtime_type:
          type: string
          enum: [agent, workflow, http, local_task, media, none]
          default: agent
        external_resource_id: { type: string, maxLength: 255 }
        identity_mode: { type: string, enum: [user, tenant], default: user }
        execution_mode: { type: string, enum: [interactive, background], default: interactive }
        session_policy: { type: string, enum: [lazy, eager], default: lazy }
        artifact_policy:
          type: string
          enum: [external_refresh, mirror_on_access, mirror_on_complete]
          default: external_refresh
        timeout_seconds: { type: integer, minimum: 10, maximum: 3600, default: 300 }
        config:
          type: object
          additionalProperties: true
    ApplicationCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100, minLength: 1 }
        slug:
          type: string
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description: { type: string }
        icon: { type: string, maxLength: 50 }
        color: { type: string, maxLength: 20 }
        category_slug: { type: string, maxLength: 100 }
        category_name: { type: string, maxLength: 100 }
        is_public: { type: boolean, default: false }
        kind: { type: string, enum: [chat, page, form, dashboard, custom, task], default: chat }
        renderer_key: { type: string, default: chat }
        runtime: { $ref: "#/components/schemas/RuntimeBindingInput" }
        set_default_agent: { type: boolean, default: false }
        skills:
          type: array
          maxItems: 20
          description: Agent-scoped 技能配置; omit or send `[]` for none.
          items: { $ref: "#/components/schemas/ApplicationSkill" }
...